Add initial core application features
This commit introduces the foundational features for the snippet management application, covering snippet lifecycle, code parsing, authentication, and a robust search system.

Key features include:
*   **Snippet Management**: Full CRUD operations for snippets, pagination, and draft status.
*   **Java Code Parsing**: Asynchronous parsing, automatic metadata extraction, fragment wrapping, and virtual thread support.
*   **Authentication & Security**: JWT-based authentication with secure HTTP-only cookies, token refresh, Redis caching, and role-based access control.
*   **Comprehensive Search**: Advanced search by tags, names, languages, and code structure, supporting both direct and fuzzy matches.
*   **Architectural Improvements**: Refactor authentication logic, standardize transactional annotations, update DDL strategy, integrate Spring Cache, and introduce custom exceptions.
*   Add H2 database for testing and enable its console.
*   Update `CHANGELOGS.md` to document these new features.

+++ src/main/java/com/victor/astronaut/snippets/SnippetMapper.java
@@ -16,8 +16,8 @@ public class SnippetMapper {
     public SnippetPreview toPreview(Snippet snippet){
         return new SnippetPreview() {
             @Override
-            public Integer getId() {
-              return Math.toIntExact(snippet.getId());
+            public Long getId() {
+              return snippet.getId();
             }
 
             @Override
